// projects/libs/repox/param-dto/src/lib/enum/param-type.enum.ts
/**
 * The enum file contains all kinds of parameter type
 * directly for the command line. It is responsible for
 * identifies all parts of command.
 */

export enum ParamTypeEnum {
  executor = "executor",
  application = "application",
  program = "program",
  command = "command",
  argument = "argument",
  alias = "alias"
}
// todo: refactor